requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
SQLAlchemy==1.4.0
async-exit-stack==1.0.1
async-generator==1.10
psycopg2-binary==2.8.6
validate-docbr==1.8.2
email-validator==1.1.2
python-jose==3.2.0
cryptography==3.4.6
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
